Split demos
