+++ samples/ControlGallery/ControlGallery/Pages/WebView/WebViewPlaygroundPage.xaml
@@ -1,0 +1,69 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
+             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+             x:Class="ControlGallery.Pages.WebView.WebViewPlaygroundPage"
+             Title="Playground">
+    <Grid Padding="20"
+          RowDefinitions="Auto,Auto,Auto,*,Auto"
+          RowSpacing="10">
+
+        <Label Grid.Row="0"
+               Text="A larger freeform WebView for trying remote URLs or packaged assets."
+               FontSize="12"
+               TextColor="Gray" />
+
+        <FlexLayout Grid.Row="1"
+                    Wrap="Wrap"
+                    Direction="Row"
+                    AlignItems="Center"
+                    JustifyContent="Start">
+            <Button x:Name="PlaygroundBackButton"
+                    Text="Back"
+                    Margin="0,0,10,10"
+                    IsEnabled="{Binding Source={x:Reference PlaygroundWebView}, Path=CanGoBack}"
+                    Clicked="OnPlaygroundBackClicked" />
+            <Button x:Name="PlaygroundForwardButton"
+                    Text="Forward"
+                    Margin="0,0,10,10"
+                    IsEnabled="{Binding Source={x:Reference PlaygroundWebView}, Path=CanGoForward}"
+                    Clicked="OnPlaygroundForwardClicked" />
+            <Button Text="Reload"
+                    Margin="0,0,10,10"
+                    Clicked="OnPlaygroundReloadClicked" />
+            <Button Text="Avalonia"
+                    Margin="0,0,10,10"
+                    Clicked="OnPlaygroundAvaloniaClicked" />
+            <Button Text="MAUI"
+                    Margin="0,0,10,10"
+                    Clicked="OnPlaygroundMauiClicked" />
+            <Button Text="Local Asset"
+                    Margin="0,0,0,10"
+                    Clicked="OnPlaygroundLocalClicked" />
+        </FlexLayout>
+
+        <Grid Grid.Row="2"
+              ColumnDefinitions="*,Auto"
+              ColumnSpacing="10">
+            <Entry x:Name="PlaygroundUrlEntry"
+                   Grid.Column="0"
+                   Placeholder="Enter URL or packaged asset path..."
+                   Text="https://avaloniaui.net"
+                   ReturnType="Go"
+                   Completed="OnPlaygroundUrlEntryCompleted" />
+            <Button Text="Go"
+                    Grid.Column="1"
+                    Clicked="OnPlaygroundGoClicked" />
+        </Grid>
+
+        <WebView Grid.Row="3"
+                 x:Name="PlaygroundWebView"
+                 Source="https://avaloniaui.net"
+                 Navigated="OnPlaygroundWebViewNavigated" />
+
+        <Label Grid.Row="4"
+               x:Name="PlaygroundStatusLabel"
+               FontSize="12"
+               TextColor="Gray" />
+
+    </Grid>
+</ContentPage>
